requirements.txt
# ── Django Core ───────────────────────────────────────────────────────────────
Django>=5.2,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
django-filter>=24.2

# ── Autenticação JWT ──────────────────────────────────────────────────────────
djangorestframework-simplejwt>=5.3

# ── Autenticação Social (Google OAuth) + Recuperação de Senha ─────────────────
django-allauth[socialaccount]>=0.56.0

# ── Banco de Dados (PostgreSQL) ───────────────────────────────────────────────
psycopg2-binary>=2.9
dj-database-url>=2.1      # lê DATABASE_URL injetada pelo Railway/DigitalOcean

# ── Segurança ─────────────────────────────────────────────────────────────────
django-axes>=6.4          # rate limiting / bloqueio por tentativas de login
whitenoise>=6.6           # serve arquivos estáticos em produção sem nginx extra

# ── Servidor de Produção ──────────────────────────────────────────────────────
gunicorn>=21.2

# ── Monitoramento de Erros (recomendado em produção) ─────────────────────────
sentry-sdk[django]>=2.0

# ── SMS via Twilio (descomente quando for ativar verificação por celular) ──────
# twilio>=9.0

# ── Variáveis de Ambiente ─────────────────────────────────────────────────────
python-dotenv>=1.0

# ── Upload e Processamento de Imagens ────────────────────────────────────────
Pillow>=10.3

# ── Armazenamento S3 / MinIO (opcional) ──────────────────────────────────────
# django-storages[s3]>=1.14
# boto3>=1.34

# ── Tarefas Assíncronas (opcional — Redis deve estar instalado) ───────────────
# celery>=5.4
# redis>=5.0
# django-celery-results>=2.5

# ── Documentação automática da API ────────────────────────────────────────────
# drf-spectacular>=0.27

# ── Utilitários de desenvolvimento ────────────────────────────────────────────
django-extensions>=3.2
